Add support to non-canonical coin systems
Before the change only canonical systems were guaranteed to work.
With this update the concept of CoinSystem is introduced.
Coin systems can be canonical, non-canonical or auto-detected (the
detection algorithm can be expensive).

For canonical systems the change-making problem can be solved
efficiently with a greedy algorithm. For non-canonical systems a
pseudo-polynomial time dynamic programming solution becomes necessary.